Split Single-life
Single-life has been split and compartmentalised into different files, this makes it so that adding features later can be easier.
